git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@117227 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/acssampGUI/src/cl/utfsm/samplingSystemUI/IGraphicalUpdater.java
@@ -13,5 +13,5 @@ public interface IGraphicalUpdater {
 	
 	public void resetSampleCount();
 	
-	public void setTimeWindow(long frecuency, int time);
+	public void setTimeWindow(double frecuency, int time);
 }
